feature(hc): add subscriptions

+++ spec/shared/zendesk_resource.rb
@@ -23,7 +23,9 @@ shared_examples 'zendesk#resource' do |options = {}|
     if options.fetch(:paged, true)
       context 'paging' do
         before(:each) do
-          @resources = Array.new(3) { collection.create!(instance_exec(&options[:create_params])) }
+          if collection.reload.count < 3
+            @resources = Array.new(3) { collection.create!(instance_exec(&options[:create_params])) }
+          end
         end
 
         after(:each) { (@resources || []).each(&:destroy) }
@@ -77,13 +79,15 @@ shared_examples 'zendesk#resource' do |options = {}|
       end
     end
 
-    it 'should be destroyed' do
-      @record = collection.create!(create_params)
-      expect(record.identity).not_to be_nil
-      record.destroy
+    if options.fetch(:destroy, true)
+      it 'should be destroyed' do
+        @record = collection.create!(create_params)
+        expect(record.identity).not_to be_nil
+        record.destroy
 
-      if !options.fetch(:delayed_destroy, false) && !Zendesk2.mocking?
-        expect(record).to be_destroyed
+        if !options.fetch(:delayed_destroy, false) && !Zendesk2.mocking?
+          expect(record).to be_destroyed
+        end
       end
     end
